[feat] 일정수정, 리뷰작성 & 리팩토링, 버그수정

+++ routy-be/src/main/java/com/c4/routy/domain/region/entity/RegionEntity.java
@@ -16,7 +16,7 @@ public class RegionEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "region_id")
-    private Long regionId;
+    private Integer regionId;
 
     @Column(name = "region_name", nullable = false)
     private String regionName;
